add option to convert patches to code
afaik, the only way to get code out of UndertaleModTool is to manually type out the names of every single file you want.

i didnt wanna do that, so i made it so if u put .gml.patch files in the Source/Code directory, then it will load the modded code and save it over the patch.